First step to refactore most of the code
A good amount of work went into rethinking the design of the software,
which in turns forced me to rewrite nearly every part of it.

With this first commit the UI changed drastically, and the way archives
are being managed switched from using a python library to using an
external tool (7zip).

More to come.