Add a filesystem watchdog over the local repository
The watchdog will monitor any change to the local repository and
automatically add or remove archives who gets added or removed from the
folder through a third party action (e.g. download from web browser).

To reduce operations, the application will not do anything unless in
focus. If changes are made while the application isn't active, a full
refresh of the managed archives is done as soon as the app regain focus.
New archives are added to the list, and missing archives removed from
the list. Doing so will not uninstall any file from the game's modules
folder.