Add watchdog for game's module repository
This is a follow up to 40c7bc0023. I had to rewrite quite a lot of code
and change how ListRowItem are managed. Each archive now has its own
object instead of being coopted to the Qt row object. An EventFilter at
the QApplication level was needed to make sure the application did indeed
become innactive.

The watchdog gets unscheduled whenever something changes one the hard
drive. This behavior needs to be replicated for the archive repository.