Make sure path ending with 'res' not be trouble
This methods make use of qmm.settings.IsDirValidator when loading path
from the setting file. It ensure the path value is a pathlib object,
which give us access to PurePath.relative_to(). That last method is a
cleaner way to properly strip the game path from whichever path needs to
be trimmed.