Fixes #55: explain why a mod can't be installed
The original problem was that a module wouldn't be installed, creating
confusion in the user as to why. The provided example, in the issue,
possess an erroneous structure which is being ignored by the software.
Sadly, the user didn't seem to understand why it was so.

This isn't a perfect resolution by far. There are currently way too many
dialog that can appear for any given problem. More over, if the "empty"
module is part of a group selection, the dialog will not appear.